requirements.txt
# Core dependencies
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0

# Text processing and NLP
emoji==2.8.0
langdetect==1.0.9
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Geocoding and location services
geopy==2.4.0

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Machine Learning for hazard detection
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2

# Social media APIs
praw==7.7.1  # Reddit API
instaloader==4.10.3  # Instagram scraping

# Utilities
python-dateutil==2.8.2
pytz==2023.3
